Guard React access in split quiz components
