+++ src/index.js
@@ -1,5 +1,4 @@
-import PlantQuizGame from './components/PlantQuizGame.js';
+import PlantQuizGame from './components/PlantQuizGame.jsx';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(React.createElement(PlantQuizGame));
-
